added authorization for goals

+++ backend/controllers/goalController.js
@@ -1,9 +1,10 @@
 const asyncHandler = require('express-async-handler');
 
 const Goal = require('../models/goalModel');
+const User = require('../models/userModel');
 
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find({});
+    const goals = await Goal.find({ user: req.user._id });
     res.status(200).json(goals);
 });
 
@@ -13,7 +14,8 @@ const setGoal = asyncHandler(async (req, res) => {
         throw new Error('Please add text field');
     }
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user._id,
     });
 
     res.status(200).json(goal);
@@ -25,6 +27,18 @@ const updateGoal = asyncHandler(async (req, res) => {
         res.status(404);
         throw new Error('Goal not found');
     }
+
+    const user = await User.findById(req.user._id);
+    if (!user) {
+        res.status(404);
+        throw new Error('User not found');
+    }
+
+    if (goal.user.toString() !== req.user._id.toString()) {
+        res.status(401);
+        throw new Error('Not authorized');
+    }
+
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
         new: true
     })
@@ -37,6 +51,18 @@ const deleteGoal = asyncHandler(async (req, res) => {
         res.status(404);
         throw new Error('Goal not found');
     }
+
+    const user = await User.findById(req.user._id);
+    if (!user) {
+        res.status(404);
+        throw new Error('User not found');
+    }
+
+    if (goal.user.toString() !== req.user._id.toString()) {
+        res.status(401);
+        throw new Error('Not authorized');
+    }
+
     await Goal.deleteOne({ _id: req.params.id });
     res.status(200).json({ message: `Deleted goal ${req.params.id}` });
 });
